# airbyte-api/src/main/openapi/api.yaml
openapi: 3.0.0
info:
  title: Airbyte Public API
  version: 1.0.0
servers:
  - url: "http://localhost:8006/v1"
    description: Airbyte API v1
paths:
...
  /destinations:
    get:
      tags:
        - Destinations
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationsResponse"
          description: Successful operation
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: listDestinations
      summary: List destinations
      parameters:
        - name: workspaceIds
          description: The UUIDs of the workspaces you wish to list destinations for. Empty list will retrieve all allowed workspaces.
          schema:
            type: array
            items:
              format: uuid
              type: string
          in: query
          required: false
        - name: includeDeleted
          description: Include deleted destinations in the returned results.
          schema:
            default: false
            type: boolean
          in: query
          required: false
        - name: limit
          description: Set the limit on the number of destinations returned. The default is 20.
          schema:
            format: int32
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          in: query
        - name: offset
          description: Set the offset to start at when returning destinations. The default is 0
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          in: query
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationCreateRequest"
            examples:
              Destination Creation Request Example:
                value:
                  name: Postgres
                  workspaceId: 2155ae5a-de39-4808-af6a-16fe7b8b4ed2
                  configuration:
                    airbyte_destination_name: postgres
                    port: 5432
                    schema: public
                    ssl_mode:
                      mode: prefer
                    tunnel_method:
                      tunnel_method: NO_TUNNEL
                    host: localhost
                    database: postgres
                    username: postgres
                    password: test
      tags:
        - Destinations
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationResponse"
              examples:
                Destination Creation Response Example:
                  value:
                    destinationId: af0c3c67-aa61-419f-8922-95b0bf840e86
          description: Successful operation
        "400":
          description: Invalid data
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: createDestination
      summary: Create a destination
      description: "Creates a destination given a name, workspace id, and a json blob containing the configuration for the source."
  /destinations/{destinationId}:
    get:
      tags:
        - Destinations
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationResponse"
              examples:
                Destination Get Response Example:
                  value:
                    destinationId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
                    name: My Destination
                    sourceType: postgres
                    workspaceId: 744cc0ed-7f05-4949-9e60-2a814f90c035
                    configuration:
                      conversion_window_days: 14
                      customer_id: "1234567890"
                      start_date: 2023-01-01
                      end_date: 2024-01-01
          description: Get a Destination by the id in the path.
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: getDestination
      summary: Get Destination details
    delete:
      tags:
        - Destinations
      responses:
        "204":
          description: The resource was deleted successfully
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: deleteDestination
      summary: Delete a Destination
    patch:
      tags:
        - Destinations
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationPatchRequest"
            examples:
              Destination Update Request Example:
                value:
                  configuration:
                    conversion_window_days: 14
                    customer_id: "1234567890"
                    start_date: 2023-01-01
                    end_date: 2024-01-01
                  name: My Destination
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationResponse"
              examples:
                Destination Update Response Example:
                  value:
                    destinationId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
                    name: running
                    sourceType: postgres
                    workspaceId: 744cc0ed-7f05-4949-9e60-2a814f90c035
                    configuration:
                      conversion_window_days: 14
                      customer_id: "1234567890"
                      start_date: 2023-01-01
                      end_date: 2024-01-01
          description: Update a Destination
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: patchDestination
      summary: Update a Destination
    put:
      tags:
        - Destinations
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationPutRequest"
            examples:
              Destination Update Request Example:
                value:
                  configuration:
                    conversion_window_days: 14
                    customer_id: "1234567890"
                    start_date: 2023-01-01
                    end_date: 2024-01-01
                  name: My Destination
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationResponse"
              examples:
                Destination Update Response Example:
                  value:
                    destinationId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
                    name: running
                    sourceType: postgres
                    workspaceId: 744cc0ed-7f05-4949-9e60-2a814f90c035
                    configuration:
                      conversion_window_days: 14
                      customer_id: "1234567890"
                      start_date: 2023-01-01
                      end_date: 2024-01-01
          description: Update a Destination and fully overwrite it
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: putDestination
      summary: Update a Destination and fully overwrite it
    parameters:
      - name: destinationId
        schema:
          format: UUID
          type: string
        in: path
        required: true
...
components:
...
  schemas:
...
    DestinationConfiguration:
      description: The values required to configure the destination.
...
    DestinationResponse:
      title: Root Type for DestinationResponse
      description: Provides details of a single destination.
      type: object
      required:
        - destinationId
        - name
        - destinationType
        - workspaceId
        - configuration
      properties:
        destinationId:
          format: UUID
          type: string
        name:
          type: string
        destinationType:
          type: string
        workspaceId:
          format: UUID
          type: string
        configuration:
          $ref: "#/components/schemas/DestinationConfiguration"
      example:
        destinationId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
        name: Analytics Team Postgres
        destinationType: postgres
        workspaceId: 871d9b60-11d1-44cb-8c92-c246d53bf87e
...
    DestinationCreateRequest:
      required:
        - name
        - workspaceId
        - configuration
      type: object
      properties:
        name:
          description: Name of the destination e.g. dev-mysql-instance.
          type: string
        definitionId:
          description: The UUID of the connector definition. One of configuration.destinationType or definitionId must be provided.
          format: uuid
          type: string
        workspaceId:
          format: uuid
          type: string
        configuration:
          $ref: "#/components/schemas/DestinationConfiguration"
      x-implements: io.airbyte.api.common.ConfigurableActor
    DestinationPatchRequest:
      type: object
      properties:
        name:
          type: string
        configuration:
          $ref: "#/components/schemas/DestinationConfiguration"
      x-implements: io.airbyte.api.common.ConfigurableActor
    DestinationPutRequest:
      required:
        - name
        - configuration
      type: object
      properties:
        name:
          type: string
        configuration:
          $ref: "#/components/schemas/DestinationConfiguration"
      x-implements: io.airbyte.api.common.ConfigurableActor
...
    DestinationsResponse:
      title: Root Type for DestinationsResponse
      description: ""
      required:
        - data
      type: object
      properties:
        previous:
          type: string
        next:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/DestinationResponse"
      example:
        next: "https://api.airbyte.com/v1/destinations?limit=5&offset=10"
        previous: "https://api.airbyte.com/v1/destinations?limit=5&offset=0"
        data:
          destinationId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
          name: Analytics Team Postgres
          destinationType: postgres
          workspaceId: 871d9b60-11d1-44cb-8c92-c246d53bf87e
...